# app/static/swagger.json
{
    "swagger": "2.0",
    "info": {
      "title": "Authors REST API",
      "version": "1.0.0",
      "description": "A simple CRUD API to manage authors"
    },
    "host": "localhost:5000",
    "basePath": "/",
    "schemes": ["http"],
    "consumes": ["application/json"],
    "produces": ["application/json"],
    "paths": {
      "/authors": {
        "get": {
          "summary": "Get all authors",
          "responses": {
            "200": {
              "description": "List of authors",
              "schema": {
                "type": "array",
                "items": { "$ref": "#/definitions/Author" }
              }
            }
          }
        },
        "post": {
          "summary": "Create a new author",
          "parameters": [
            {
              "in": "body",
              "name": "author",
              "description": "Author object to create",
              "required": true,
              "schema": { "$ref": "#/definitions/NewAuthor" }
            }
          ],
          "responses": {
            "201": {
              "description": "Author created successfully",
              "schema": { "$ref": "#/definitions/Author" }
            },
            "400": { "description": "Invalid input, name is required" }
          }
        }
      },
      "/authors/{id}": {
        "get": {
          "summary": "Get an author by ID",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "ID of author to fetch",
              "required": true,
              "type": "integer",
              "format": "int64"
            }
          ],
          "responses": {
            "200": {
              "description": "Author details",
              "schema": { "$ref": "#/definitions/Author" }
            },
            "404": { "description": "Author not found" }
          }
        },
        "put": {
          "summary": "Update an existing author",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "ID of author to update",
              "required": true,
              "type": "integer",
              "format": "int64"
            },
            {
              "in": "body",
              "name": "author",
              "description": "Author object with updated info",
              "required": true,
              "schema": { "$ref": "#/definitions/NewAuthor" }
            }
          ],
          "responses": {
            "200": {
              "description": "Author updated",
              "schema": { "$ref": "#/definitions/Author" }
            },
            "404": { "description": "Author not found" }
          }
        },
        "delete": {
          "summary": "Delete an author",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "ID of author to delete",
              "required": true,
              "type": "integer",
              "format": "int64"
            }
          ],
          "responses": {
            "200": { "description": "Author deleted" },
            "404": { "description": "Author not found" }
          }
        }
      }
    },
    "definitions": {
      "Author": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "name": { "type": "string" },
          "bio": { "type": "string" }
        },
        "required": ["id", "name"]
      },
      "NewAuthor": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "bio": { "type": "string" }
        },
        "required": ["name"]
      }
    }
  }
  